requirements.txt
streamlit
spacy
nltk
sentence-transformers
PyPDF2
python-docx
scikit-learn
matplotlib
plotly
seaborn
